ESP8266: strip (optional) trailing \r from received serial lines
The lines are assumed to be delimited by \n (which is effectively
stripped). Now, we also check the preceding character and strip
it if it is '\r'. This allows us to handle both LF and CRLF newlines
in incoming serial communication. (Previously, the \r would remain
at the end of the lines).